Extract common game logic.
