+++ front/src/common/question.scss
@@ -1,47 +1,47 @@
 .question {
-    text-align: center;
+  text-align: center;
 
-    .operation {
-        font-size: 52px;
-        font-weight: bold;
-        background: hsl(145 100% 81%);
-        border-radius: 15px;
-        width: 50%;
-        margin: auto;
+  .operation {
+    font-size: 52px;
+    font-weight: bold;
+    background: hsl(145 100% 81%);
+    border-radius: 15px;
+    width: 50%;
+    margin: auto;
+  }
+
+  .answers {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    grid-template-rows: repeat(2, 5em);
+    margin: 1em;
+    margin-top: 3em;
+
+    > * {
+      padding: 0.5em;
     }
 
-    .answers {
-        display: grid;
-        grid-template-columns: repeat(3, 1fr);
-        grid-template-rows: repeat(2, 5em);
-        margin: 1em;
-        margin-top: 3em;
-
-        >* {
-            padding: 0.5em;
-        }
-
-        button {
-            font-size: 26px;
-            font-weight: bold;
-            border-radius: 10px;
-            box-shadow: 0 4px 6px rgba(88, 51, 223, 0.3);
-            background: linear-gradient(135deg, #027bb8, #075994);
-            width: 100%;
-            height: 100%;
-            color: white;
-        }
-
-        .correct {
-            background: linear-gradient(135deg, #027408, #51e471);
-        }
-
-        .invalid {
-            background: linear-gradient(135deg, #df0707fd, #e19d9dd9);
-        }
-
-        :disabled {
-            background: linear-gradient(135deg, #2b2c2c, #bfcad1);
-        }
+    button {
+      font-size: 26px;
+      font-weight: bold;
+      border-radius: 10px;
+      box-shadow: 0 4px 6px rgba(88, 51, 223, 0.3);
+      background: linear-gradient(135deg, #027bb8, #075994);
+      width: 100%;
+      height: 100%;
+      color: white;
     }
+
+    .correct {
+      background: linear-gradient(135deg, #027408, #51e471);
+    }
+
+    .invalid {
+      background: linear-gradient(135deg, #df0707fd, #e19d9dd9);
+    }
+
+    :disabled {
+      background: linear-gradient(135deg, #2b2c2c, #bfcad1);
+    }
+  }
 }
