+++ front/src/multiplication/game_settings.scss
@@ -1,31 +1,31 @@
 .game-settings {
-    padding: 3em 1em;
-    display: grid;
-    grid-template-columns: 1fr 2fr;
-    row-gap: 2em;
-    select.difficulty-level {
-        width: 90%;
-    }
+  padding: 3em 1em;
+  display: grid;
+  grid-template-columns: 1fr 2fr;
+  row-gap: 2em;
+  select.difficulty-level {
+    width: 90%;
+  }
 
-    .buttons {
-        grid-column-start: 1;
-        grid-column-end: 3;
-        text-align: center;
+  .buttons {
+    grid-column-start: 1;
+    grid-column-end: 3;
+    text-align: center;
 
-        button {
-            background: linear-gradient(135deg, #4299e1, #63b3ed);
-            color: white;
-            font-size: 22px;
-            font-weight: bold;
-            padding: 16px 10px;
-            border-radius: 12px;
-            border: none;
-            cursor: pointer;
-            box-shadow: 0 4px 6px rgba(66, 153, 225, 0.3);
-            transition: all 0.2s ease;
-            text-align: center;
-            position: relative;
-            overflow: hidden;
-        }
+    button {
+      background: linear-gradient(135deg, #4299e1, #63b3ed);
+      color: white;
+      font-size: 22px;
+      font-weight: bold;
+      padding: 16px 10px;
+      border-radius: 12px;
+      border: none;
+      cursor: pointer;
+      box-shadow: 0 4px 6px rgba(66, 153, 225, 0.3);
+      transition: all 0.2s ease;
+      text-align: center;
+      position: relative;
+      overflow: hidden;
     }
+  }
 }
