+++ front/src/multiplication/multiplication.scss
@@ -1,17 +1,17 @@
 .game-board {
-    .lifes-display {
-        position: absolute;
-        top: 15px;
-        left: 20px;
-    }
+  .lifes-display {
+    position: absolute;
+    top: 15px;
+    left: 20px;
+  }
 
-    .remaining-questions {
-        position: absolute;
-        top: 15px;
-        right: 15px;
-    }
+  .remaining-questions {
+    position: absolute;
+    top: 15px;
+    right: 15px;
+  }
 
-    .question {
-        margin-top: 50px;
-    }
+  .question {
+    margin-top: 50px;
+  }
 }
